+++ app/src/main/java/com/nikola/jakshic/dagger/vo/Peer.kt
@@ -11,5 +11,5 @@ data class Peer(
         @ColumnInfo(name = "peer_id") @SerializedName("account_id") @Expose var peerId: Long,
         @ColumnInfo(name = "persona_name") @SerializedName("personaname") @Expose var personaname: String?,
         @ColumnInfo(name = "avatar_url") @SerializedName("avatarfull") @Expose var avatarfull: String?,
-        @ColumnInfo(name = "games") @SerializedName("with_games") @Expose var withGames: Long,
-        @ColumnInfo(name = "wins") @SerializedName("with_win") @Expose var withWin: Long)
+        @ColumnInfo(name = "games") @SerializedName("with_games") @Expose var withGames: Int,
+        @ColumnInfo(name = "wins") @SerializedName("with_win") @Expose var withWin: Int)
